+++ BambooTracker/gui/pattern_editor/pattern_editor_panel.hpp
@@ -87,6 +87,7 @@ signals:
 
 	void selected(bool isSelected);
 	void returnPressed();
+	void instrumentEntered(int num);
 
 protected:
 	virtual bool event(QEvent *event) override;